// lib/data/types.ts
export interface Product {
  id: number
  label: string
  description: string
  img: string
  cuidados: string
  cantidad: number
  precio: number
}

export interface Design {
  id: number
  label: string
  description: string
  img: string
  materiales: string
  cantidad: number
  precio: number
}